+++ src/utils/schema.ts
@@ -5,9 +5,36 @@ import type z from 'zod';
 import {zodToJsonSchema} from 'zod-to-json-schema';
 import type {GenContext} from '../gen/gen';
 
+/**
+ * Vocab schemas have `$anchor` instead of `$id`,
+ * and are expected to be transformed or bundled with `bundleSchemas`.
+ */
 export interface VocabSchema extends JSONSchema {
-	$id: string;
+	$anchor: string;
 }
+
+/**
+ * Bundles an array of `VocabSchema`s into a single spec-compliant `JSONSchema`.
+ * @param schemas
+ * @param $id - @example 'https://app.felt.dev/schemas/vocab.json'
+ * @param title - @example '@feltjs/felt-server vocab'
+ * @param $schema - Defaults to version 2020-12.
+ * @returns
+ */
+export const bundleSchemas = (
+	schemas: VocabSchema[],
+	title: string,
+	$id: string,
+	$schema = 'https://json-schema.org/draft/2020-12/schema',
+): JSONSchema => {
+	const schema: JSONSchema = {
+		$id,
+		$schema,
+		title,
+		$defs: structuredClone(schemas).sort((a, b) => a.$anchor.localeCompare(b.$anchor)),
+	};
+	return schema;
+};
 
 export const isVocabSchema = (value: unknown): value is VocabSchema =>
 	!!value && typeof value === 'object' && '$id' in value;
@@ -15,10 +42,12 @@ export const isVocabSchema = (value: unknown): value is VocabSchema =>
 export const toVocabSchema = (t: z.ZodType<any, z.ZodTypeDef, any>, name: string): VocabSchema => {
 	const schema = zodToJsonSchema(t, name);
 	const args = (schema.definitions ? schema.definitions[name] : {}) as VocabSchema;
-	args.$id = `/schemas/${name}.json`;
+	args.$anchor = '#' + name;
 	return args;
 };
 
+// TODO BLOCK remove all /schemas/ and $id usage
+// TODO BLOCK try to delete this resolver
 /**
  * Creates a custom resolver for `VocabSchema`s supporting paths like "/schemas/Something.json".
  * @param schemas