+++ changelog.md
@@ -1,5 +1,11 @@
 # changelog
 
+## 0.76.0
+
+- **break**: replace `$id` on `VocabSchema` with `$anchor`,
+  and add the `bundleSchemas` helper
+  ([#372](https://github.com/feltjs/gro/pull/372))
+
 ## 0.75.5
 
 - upgrade `@ryanatkn/json-schema-to-typescript` dep,